WIP Add text input SCSS module
