+++ web/src/components/button/button.module.scss
@@ -1,4 +1,4 @@
-@import '@/scss/colors.scss';
+@import '@/scss/colors';
 
 .btn {
   padding: 8px 16px;
